rename type_test to type_testing
