+++ cupy/_logic/type_testing.py
